requirements.txt
django==3.2.16
django-phone-field==1.8.1